List of all functions

Here an overview of all functions:

JustRelax.Geometry — Type

struct Geometry{nDim,T}

A struct representing the geometry of a topological object in nDim dimensions.

Arguments

nDim: The number of dimensions of the topological object.

T: The type of the elements in the topological object.

JustRelax.cart2ind — Method

cart2ind(A)

Return the linear index of a n-dimensional array corresponding to the cartesian indices I

JustRelax.element — Method

element(A, element_indices..., cell_indices...)

Return a the element with element_indices of the Cell with cell_indices of the CellArray A.

Arguments

element_indices::Int|NTuple{N,Int}: the element_indices that designate the field in accordance with A's cell type.

cell_indices::Int|NTuple{N,Int}: the cell_indices that designate the cell in accordance with A's cell type.

JustRelax.setelement! — Method

setelement!(A, x, element_indices..., cell_indices...)

Store the given value x at the given element with element_indices of the cell with the indices cell_indices

Arguments

x::Number: value to be stored in the index element_indices of the cell with cell_indices.

element_indices::Int|NTuple{N,Int}: the element_indices that designate the field in accordance with A's cell type.

cell_indices::Int|NTuple{N,Int}: the cell_indices that designate the cell in accordance with A's cell type.

JustRelax.velocity_grids — Method

velocity_grids(xci, xvi, di::NTuple{N,T}) where {N,T}

Compute the velocity grids for N dimensionional problems.

Arguments

xci: The x-coordinate of the cell centers.

xvi: The x-coordinate of the cell vertices.

di: A tuple containing the cell dimensions.